Update styles css
