+++ package.json
@@ -21,6 +21,7 @@
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
     "redux": "^4.2.0",
+    "sass": "^1.56.2",
     "slick-carousel": "^1.8.1",
     "web-vitals": "^2.1.4"
   },
