+++ src/styles/carrocel.css
@@ -18,4 +18,3 @@
   height: 10vw;
   object-fit: cover;
 }
-
